Multi-Realm Cleanup
- Refactored `getEnabledRealms` from `acl-config` and `DatawarehouseInitializer`
  to `Realms` and replaced the duplicate code with calls to the newly refactored
  function.
- Updated `Users::getCentersFor` to return an array containing the user's
  organizationId if no records were found in the DB.
- Utilized the new `Realms::getEnabledRealms` to update the dreaded
  `summary3.php`. This meant that we could remove the `.diff` files & update
  `bootstrap.sh` to not apply `.diff`.
